Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> What's the difference between registers-common.h and registers-all.h?

In this patch, registers-all.h is a super-set of registers-common.h.
The difference is that registers-common.h contains the registers that are common between the STM32H7 family (now registers-stm32h7.h) and all others (registers-all.h).

This is a setup for the next patch, which factors out the overlap.

Patch-set: 2
Reviewer: Gerrit User 1334530 <1334530@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>