elm/hana: Set accel min frequency to 12.5Hz
The EC is running at 10Hz itself, so it won't be able to properly
handle frequencies below that threshold, and data will end up being
sent to the kernel at 10Hz.

This was confusing CTS as it expected to be able to program the
sensor to send data at 1Hz, but receiving it at 10Hz instead.

Fix the issue by picking the closest frequency above 10Hz that the
accelerometer supports, i.e. 12.5Hz

BUG=b:134422740
TEST=run CtsSensorTestCases, observe it pass

Change-Id: I4804dc114dbd11caeb2c2a6d6500bc30ed88abc9
Signed-off-by: Enrico Granata <egranata@chromium.org>