g: enforce order of enabling RO/RW sections
When a post update reset happens and the host sends the vendor command
to enable the update, in case both RO and RW have been updated, the
Cr50 should enable the RW section first, and then the RO. This would
cover the case when the new RO has the new key and would not be able
to start the old RW, so the matching RW must be available first.

Enabling RW without enabling RO is not a problem, worst thing which
could happen is that the old RO would not start the new RW and the
update process will have to be repeated.

BRANCH=cr50, cr50-mp
BUG=b:74100307
TEST=none yet. Will verify when a new RO is available.

Change-Id: I00175a5a957166d4423fb270bbe9f92d4e408d5c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1644479
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
(cherry picked from commit 11f2760452059e9769c7500984fac6402c565ad1)