cr50: optimize set_ec_on() by prioritizing the case EC is on
This patch prioritize the case that EC is in ON over other cases like
INIT, INIT_DEBOUNCING or DEBOUNCING in set_ec_on(). set_ec_on() is
called every second to monitor EC status when EC_UART signal is
detected. set_ec_on() shall finish quickly in most of time when EC
is ON.

BUG=None
BRANCH=None
TEST=ran flash_ec and uart_stress_tester.sh on Bob.

Signed-off-by: Namyoon Woo <namyoon@chromium.org>

Change-Id: I6a53ca86ccb4f65c7450bfbfda78cb5cf5b6409d