Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I think we should hold off on this until we release 0.3.18. Other patches may land before we start 0.4.19 qual.

I am not sure I follow - whatever happens, the next pre-pvt release will have to be at least 0.4.19, right?

Or are you worried that this patch will not be the last one in the list? This is not a big deal, we do not rely on the manifest change as the marker of a new branch, we go by the sha in the version.

Patch-set: 1
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>