usb_pd: Send Type-C's USB3 and USB2 physical port mapping info
In order to configure the Intel Virual MUX, USB3 and USB2 port numbers
of the USBC port are needed. USBC port numbers may not be physically
mapped to same USB3 & USB2 port number as 1 to 1.
Based on the complexity of physical layout of the board, USB3 & USB2
port signals are routed for respective USBC port.

As EC knows the information of the USBC port number, USB3 & USB2 port
mapping also can be added in the EC and can be passed to AP upon request.

BUG=b:134614710
BRANCH=none
TEST=Manually tested on ICLRVP, able to configure the Virtual MUX
     based on the USB3 & USB2 port mapping information provided by EC.

Change-Id: I8714355197dc84ff0536e174931685e027f4d9ef
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>