usb_pd: Send Type-C's USB3 and USB2 physical port number info
On Intel devices EC's USB-C port numbers may not be physically equal to
AP's USB3 & USB2 port number. Because there can be MAX 15 USB2 ports on
PCH and MAX 15 USB3 ports on SOC, based on the complexity of the physical
layout of the board, USB3 & USB2 port signals of AP are routed to
respective USB-C port of EC. Hence, to configure the Intel Virtual MUX,
information of USB3 and USB2 port numbers of the respective USB-C port is
needed. As EC knows the information of the number of USB-C port number,
associated USB3 & USB2 port numbers of USB-C port also can be added in
the EC and can be sent to AP upon request.

BUG=b:134614710
BRANCH=none
TEST=Manually tested on ICLRVP, able to configure the Virtual MUX
     based on the USB3 & USB2 port number information provided by EC.
     On ICL-U RVP using ITE EC following are the configuration.
     -----------------------------------------------------
     EC's USB-C Port# | SOC's USB3 port# | PCH's USB2 port#
     -----------------------------------------------------
              0       |        1         |      4
     -----------------------------------------------------
              1       |        2         |      5

Change-Id: I8714355197dc84ff0536e174931685e027f4d9ef
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>