usb_pd: Send Type-C's USB3 and USB2 physical port mapping info
In order to configure the Intel Virtual MUX, USB3 and USB2 port numbers
of the USB-C port are needed. USB-C port numbers may not be physically
mapped to same USB3 & USB2 port number as 1 to 1.
There can be MAX 15 USB2 ports on PCH and MAX 15 USB3 ports on SOC.
Based on the complexity of physical layout of the board, USB3 & USB2
port signals are routed for respective USB-C port.

As EC knows the information of the USB-C port number, USB3 & USB2 port
mapping also can be added in the EC and can be passed to AP upon request.

BUG=b:134614710
BRANCH=none
TEST=Manually tested on ICLRVP, able to configure the Virtual MUX
     based on the USB3 & USB2 port mapping information provided by EC.

Change-Id: I8714355197dc84ff0536e174931685e027f4d9ef
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>