driver: set Kionix accel min frequency to 12.5Hz
The EC is running at 10Hz itself, so it won't be able to properly
handle frequencies below that threshold, and data will end up being
sent to the kernel at 10Hz.

This is a source of CTS issues on some devices, as Android expects
to be able to program the sensor to send data at 1Hz given the
sensor's configuration parameters.

Fix the issue by picking the closest frequency above 10Hz that the
accelerometer supports, i.e. 12.5Hz

Conflicts:
   MOTION_MAX_SENSOR_FREQUENCY not defined in the branch

BUG=b:134422740
BRANCH=oak
TEST=run CtsSensorTestCases, observe it pass

Change-Id: I56772009817e3cbd452e96637c0a78f54f3854c7
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1647363
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>