vprintf: make pointer to context const
Pointer to the context used by the __tx_char() callback of vprintf()
and friends is not constant, but there is no reason for it to be, as
all uses are fine with const context pointer. In fact the need to keep
the pointer not const causes some unnecessary assignments in the code
just to copy a constant context object into a non constant one.

The following patch has some example of such assignment.

BRANCH=cr50, cr50-mp
BUG=b:38448364
TEST=sizes of all ROM and RAM spaces reported by

  make buildall -j | grep 'bytes in flash and ' | sort

   before and after this patch are the same.

Change-Id: I1e7cf7368f674a97e0f3dbd987253ea866a95c1f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>