anx7447: USB Mux is reset when TCPM is reset
Calling tcpm_init will reset the Analogix USB Mux on port 0. I'm not
sure if this is the proper behavior but I lean towards that it's not
because there is a separate function, usb_mux_init, for resetting
the usb mux. Also, calling tcpm_init for the Parade does not reset
it's mux.

BUG=b:134829988
BRANCH=none
TEST=manual
  HATCH (Port 0): The original and new stack was tested as follows:
                  1) Plug in dock
                  2) Depending on dock, plug in HDMI or DP
                  3) Plug TypeC power into dock
                  4) Verify projection on external monitor isn't lost

Change-Id: I88d17479f4d5810be3686cd78545cc99b0c41347
Signed-off-by: Sam Hurst <shurst@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1650728
Tested-by: Sam Hurst <shurst@google.com>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Commit-Queue: Sam Hurst <shurst@google.com>