Kefka: Wake up from hibernation periodically
This patch makes EC wake up from hibernation every 23 hours. EC will go
back to hibernation after it's left idle for 60 minutes. This cycle will
be repeated until a battery reaches critical level. When battery reaches
the critical level, EC cuts off a battery.

If a charger is plugged, EC stays in G3. This is not a new behavior.
If a power button is pressed, EC should continue to boot.

This patch also
- Enables battery cutoff on critical charge.
- Sets critical charge level to 6% (5% to cutoff)
- Disables pseudo G3 (to enable chip hibernation)

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/128455667,b/134973675
BRANCH=strago
TEST=Verify sabin repeats hibernate and wake-up periodically.
TEST=Verify sabin boots from hibernation when a power button is pressed.

Change-Id: I472c406bd709a1ed2e24f0923360ec4ad239ec26