docs: Update README
Converted the documentation in
http://dev.chromium.org/chromium-os/ec-development to Markdown with
go/gdc. I made minor cleanups to the Markdown after the conversion due
to bugs in the formatting, but preserved the content/text. It's known
that this documentation is out of date, but I wanted to preserve the
history, so updates to the context/text should be made after this
commit.

BRANCH=none
BUG=chromium:973205
TEST=make buildall -j
TEST=View Markdown Preview in CLion
TEST=View Markdown in gitiles

Change-Id: Ifdb2494f0f82b7226c1c77925e6a03656e604d14
Signed-off-by: Tom Hughes <tomhughes@chromium.org>