fpsensor: Prepare derivation of positive match secret and validation value.
Implement derivation of positive match secret for a template. This
derivation will be used when biod sends a command to ask for
positive match secret and the secret readable bit is set. Also add
unit tests for this derivation. Also add derivation of validation
value, which is hmac-SHA256 of positive match secret. In the case
of enrollment, the validation value will be sent to biod together
with the template.

BRANCH=nocturne
BUG=chromium:927095
TEST=make -j buildall
TEST=tested enrollment, matching and multifinger on DUT nocturne

Change-Id: Ife477a98573284f69c5fb07e814d9a3d09f92127
Signed-off-by: Yicheng Li <yichengli@chromium.org>