fpsensor: Implement derivation of positive match secret.
Implement derivation of positive match secret given a positive macth
salt and a finger id. This derivation will be used when biod sends a
command to ask for positive match secret and the secret readable bit
is set. Also add unit tests for this derivation.

BRANCH=nocturne
BUG=chromium:927095
TEST=ran unittests
TEST=tested enrollment, matching and multifinger on DUT nocturne
TEST=verified positive match secret, positive match salt and finger
id for enrolled fingers live properly in memory by printing to log
(the printing statements are deleted before commit).

Change-Id: Ife477a98573284f69c5fb07e814d9a3d09f92127
Signed-off-by: Yicheng Li <yichengli@chromium.org>