Update patch set 1
Patch Set 1:

(1 comment)

My reading of the spec is that we shouldn't be retrying hard resets forever, but only nHardResetCount times (=2 in PD 2.0). If that still fails, ErrorRecovery should be performed. Additionally, the HardResetComplete timer seems to be a maximum timeout value which after it expires, the protocol layer should assume the HardReset was sent.

Patch-set: 1