Update patch set 2
Patch Set 2:

> Patch Set 2: Verified+1 Commit-Queue+2
> 
> (3 comments)
> 
> > Patch Set 1:
> > 
> > (1 comment)
> > 
> > My reading of the spec is that we shouldn't be retrying hard resets forever, but only nHardResetCount times (=2 in PD 2.0). If that still fails, ErrorRecovery should be performed. Additionally, the HardResetComplete timer seems to be a maximum timeout value which after it expires, the protocol layer should assume the HardReset was sent.
> 
> right, bug implementing nHardResetCount is beyond the scope of this patch.
> let's get the timeout/watchdog issue taken care of first.  HardResetComplete
> seems like the closest thing to what we need here.

Agreed with Aseda, it seems like we're supposed to assume the hard reset was sent after the timer expires, rather than sending another from this state (looking at PRL_HR_Wait_for_PHY_Hard_Reset_Complete state).  Reconciling this code with the spec can be tricky, but definitely increases my appreciation for the new stack's consistent naming with it. 

We have nHardResetCount implemented in the form of PD_HARD_RESET_COUNT in our protocol file.  And the no response timer is be used to determine when to send more hard resets after the first.

Patch-set: 2