Update patch set 5
Patch Set 5:

> Patch Set 5: -Verified
> 
> > Patch Set 5: Code-Review-1
> > 
> > This is good progress, but I think you need to add more detailed decoding as there are several kinds of VDOs.
> > 
> > Passive cables:
> > PD 2.0 spec has SSTX and SSRX directionality.
> > PD 3.0 spec removes directionality and instead includes a field for maximum vbus voltage.
> > 
> > The raw u32 will need to be interpreted differently depending on if the cable responds back as PD 2.0 or PD 3.0.
> > 
> > Active cables:
> > If we have an active cable, there are several fields that change even in the PD 2.0 spec.
> > 
> > Secondly, the PD 3.0 spec for Active Cable  includes a *secondary* u32 vdo in addition to the primary cable VDO called Active Cable VDO 2 which you should decode as well.
> > 
> > See PD 3.0 6.4.4.3.1.5 Active Cable VDOs
> 
> Okay. I can add separate cable_vdo based on PD spec 2.0, 3.0 and the type of cable

Based on my interpretation of the PD 2.0 and 3.0 specs, there are some overlaps on the actual u32s that can be returned from the cable e-marker.

All cables with e-markers will return the 1st u32, the cable emarker:

1. Passive cable PD 2.0
2. Passive cable PD 3.0
3. Active cable PD 2.0
4. Active cable PD 3.0

Each of the 4 possibilities will be interpreted slightly differently, but are fundamentally the same u32 in storage terms.

If the cable is active, and PD 3.0, then a *second* u32 called Active Cable VDO 2 is present, and must be stored and decoded separately.

By the way, Ayushee, I am planning on doing very similar decoding in some changes in the type-c connector class i am working on. I hope to reuse some of the same code you've done here.

My plan is to eventually have the cable VDOs (1 or 2 u32s) passed up from the EC to the kernel where we have the connector class to do the same decoding for the kernel's consumers.

Patch-set: 5
CC: Gerrit User 1203114 <1203114@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1003039 <1003039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>