servo_v4: Support dual-role port
Two new roles "drp" and "drpdts" is introduced to the "cc" command.

Support a new flag to enable DRP. The do_cc() function becomes having 4
arguments. So change to merge them into a single cc_config with
different bit masks.

For the force_source or force_sink role, we explicitly set the the Rp
or Rd resistors on CC lines. But for DRP role, the Rp/Rd toggling is
controlled by the PD state machine. So don't set any CC pull resistor
for DRP.

Support an "on" option in "cc" command to restore the previous role
before emulating detach.

Add "usbc_action drp" to toggle the DRP state, which is compatible with
Plankton.

BRANCH=servo
BUG=b:135691171
TEST=Ran the firmware_PDTrySrc test passed, and not break all the
existing PD FAFT tests.

Change-Id: I3b90611c9840f502e496c42f80354e7ee002f96e
Signed-off-by: Wai-Hong Tam <waihong@google.com>