Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> lgtm, but for some reason I can't see the preview of updating_and_flashing.md in gitiles. (this site can't be reached) but I can see the others. Anybody else seeing this?

Yeah, it's happening to me too. I think it might be a caching issue. In the first patchset I forgot to include the file and tried to view it. Hopefully when it gets committed it will have a different URL that will bust the cache... (it views ok locally)

Patch-set: 3
Reviewer: Gerrit User 1314845 <1314845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>