Update patch set 2
Patch Set 2:

> Patch Set 2: Verified+1
> 
> > Patch Set 1:
> > 
> > without looking at the changes yet: why would we ever want to send at any other than maximum speed?
> 
> Modified to keep the transfer rate fixed. Changed the CL text as well.
> For now the max transfer rate is 7200 char/s, which could be regarded as 57600 bps.

I think this is an indication of the fact that this test approach is not adequate. We need to be testing full speed consoles. 

In any case I can't think of a reason to limit speed when testing.

Patch-set: 2
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>