printf: Fix hexdump and string 0 precision
This patch addresses a few issues with the current formatter.
The major points are as follows:
1. Cannot specify precision 0 (truncate all) for string or hexdump
2. Forced safe precision for malformed strings
3. No padding when using hexdump
4. Bad error EC_ERROR_INVAL in vsnprintf
5. Documentation errors

For (1), no piece of code explicitly sets the precision to 0 in order to
invoke the default behavior, which is currently no precision limit.
You can check using the following grep line:
 grep -rI '%[\*0-9]\{0,20\}\.0\{1,20\}[a-zA-Z]'
However, there are many cases where the precision is used to limit
the character output (as it should be).
 grep -rI '%[\*0-9]\{0,20\}\.[\*0-9]\{1,20\}[a-zA-Z]'
There are many more instances that use variable precision without
checking if the precision is zero. One of which is the following:
 crrev.com/4a4e2c71a0f6aaa50e0728922f84a7d54c14380a/test/host_command_fuzz.c#116
 https://clusterfuzz.com/testcase-detail/5699023975088128
Our current implementation will insert ERROR and stop processing, if a precision
of zero is detected with the hexdump. This results in a badly formatter
console line or runtime string, when the intended behavior would be to
simply read no bytes.
Our printf should handle explicit zero precision similar to stdlib's printf,
which means truncating all the way to zero positions, if specified.

For (2), our current implementation uses strlen to identify the length
of the input string, regardless of the set precision.
Since this is an embedded platform, we should use strnlen to impose safe
limits, when a precision is specified.

For (3), our implementation should support padding and adjusting of
all formatter types, since that is a primary feature of a printf
formatter.

The remaining commented code highlights odd behavior that should be
fixed at some point, but is not critical.

BUG=chromium:974084
TEST=Checked for any format lines that rely on a set precision of 0
     grep -rI '%[\*0-9]\{0,20\}\.[\*0-9]\{1,20\}[a-zA-Z]'
TEST=make run-printf V=1
BRANCH=none

Change-Id: I897c53cce20a701fcbe8fb9572eb878817525cc3
Signed-off-by: Craig Hesling <hesling@chromium.org>