fpsensor: Implement command to read positive_match_secret.
Add EC command to read positive_match_secret and set
positive_match_secret_readable bit on enrollment or matching success.
If the attempt to read is 5 seconds after the enrollment or match,
the read is not allowed (the positive_match_secret_readable bit is
cleared). Test that the command can read the data correctly and can
read for each finger only once. Test that attempt to read secret
after deadline will be rejected.

BRANCH=nocturne
BUG=chromium:927095
TEST=make buildall
TEST=tested enrollment, matching and multifinger on DUT nocturne
TEST=tested that if biod requests to download template and secret
for a finger that's not currently matched, reading secret will fail.

Change-Id: Idc734c6392d271e2aaee1cddf7c2c5b81b727b4a
Signed-off-by: Yicheng Li <yichengli@chromium.org>