WIP, stats_manager: decouple printing from generating the table
This change decouples printing a pretty table from generating said
pretty table so that this logic can be used in the upcoming power-tree
code as well. The motivation to allow for configurable alignments is
that for power-tree, to show hierarchy, we want to be able to use
indentation to show those hierarchies.

BUG=NONE
TEST=manual testing
(1) start servod with a power config on it
(2) dut-power -t 3
> see normal table
(2.1) - hack the code here to when calling PrettyPrint pass
align_config={'NAME':True}
      - sudo emerge ec-devutils -j32
      - dut-power -t 3
      - observe how the rails are now left-aligned, but the rest is
      still right aligned

...
@@ -------------------------------------------------------------
@@                        Onboard INA
@@ -------------------------------------------------------------
@@ NAME                    COUNT    MEAN  STDDEV     MAX     MIN
@@ AP_DVDD28_MSDC1_mw          4    0.00    0.00    0.00    0.00
@@ AP_DVDD_CORE_mw             4    0.00    0.00    0.00    0.00

Change-Id: I4281049fbfa34152a2bcca9f7b706906c0348abb
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>