intel_x86: Keep events_b upto date.
Currently on every resume (S3/S0ix) bios reads and clears all the pending
events in the process of logging the reason for wake. This is a problem
if the kernel needs to know the reason for the wake.

Instead use events_copy_b for logging the wake sources.
For this to work events_copy_b needs to upto date. i.e when the host
consumes the event, clear it from events_copy_b too. Please look at
go/hostevent-refactor for more info.

BUG=b:133262012
BRANCH=None
TEST=Hack hatch bios and make sure hostevent log is correct.

Change-Id: Id4c60b6b7ea1ecb500e76387aab59c58729e12a1
Signed-off-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>