Update patch set 1
Patch Set 1: Verified+1

> Patch Set 1: Code-Review-2
> 
> This doesn't actually fix it. The issue is in the pop function and the size of the count variable since CONFIG_ACCEL_FIFO is 256 in Arcada.
> 
> See https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1665213 as an alternative.

Checking the src and dst parameters of this memcpy has not negative effect but is the guarantee. It is not conflict with your claimed fix and should be included in the code. Without memcpy parameters checking already causes strange and hard to debug issue and we need to make sure it doesn't happen again.

Patch-set: 1
Reviewer: Gerrit User 1333246 <1333246@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1