ish: allow for latency in clock_event_get
__hw_clock_event_get returns a time value of when we expected to
wake up.  Due to latency in the code, it is possible for our
current time to be later than the expected wakeup time. This
is used to determine the level of sleep we will be going into.
Since next_delay is a signed value, this delta calculation can
result in a negative value.  Then when it is passed to pm_process
as an unsigned, it becomes very large.  The easiest way to deal
with this anamoly is to turn the negative number into 0 before
calling pm_process

BUG=b:133459192
BRANCH=none
TEST=Test on arcada platform to verify D0i2/D0i3 are not entered
for short delays

Change-Id: Ifb78e74f841561688d64c0a63c79c3e53c2ddc33
Signed-off-by: Denis Brockus <dbrockus@chromium.org>