Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I updated the master CL to increase ALERT_STORM_MAX_COUNT and reduce the likelihood of interfering with good port partners.  It has tested well with the bad combination I've been using (ANX TCPC + 7-magic hub) if you'd like to try it with the 87-Watt charger + old TCPC FW
> 
> we fixed the 87w charger related watchdog issue here:
> https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1654872/
> this one isn't strictly an interrupt storm but rather unfortunate interrupt
> handling.

Ultimately, it's up to you what you'd like to pull into the atlas branch.

For the root cause of the hard reset issue which is now fixed, does it require that the system go through a RO->RW jump?  That was mentioned as the first of the reproduction steps.  When we hit EC panics on octopus dogfood with the 87W charger, they were 30-40 minutes after boot, so I'm unclear on how they would have the same root cause if the jump is required.

Patch-set: 1
Reviewer: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>