Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > Patch Set 10:
> > 
> > > Patch Set 10:
> > > 
> > > (1 comment)
> > > 
> > > With the only purpose of last_deadline to allow the idle loop to determine the level of sleep we are going into, couldn't we leave everything as it was in hwtimer.c and change __idle to have the following?
> > > 
> > >     while (1) {
> > >         t0 = get_time();
> > >         next_delay = __hw_clock_event_get() - t0.le.lo;
> > > 
> > >         pm_process(t0, MAX(0, next_delay));
> > >     }
> > 
> > Changes in hwtimer.c is about accuracy. basically real event is off by MINIMUM_EVENT_DELAY_US(800us) + interrupt latency(miminum is 3 ticks(93us).
> 
> codes in power_mgt.c is for handling +alpha that's coming from increased interrupt latency.

Does it make any difference?  We expect to wake up at time X and the latency wakes us at X+alpha.  The take away is that the expected time may have already come and gone and thus turns negative in our calculation. This causes the signed to unsigned conversion turn into a really large unsigned value in this case.  So without having to change a lot of other code, we could just add the single MAX in the call that this anomaly happens and the issue is gone.

Patch-set: 10
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>