Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > Can we have this done in https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1643247 ?
> > > 
> > > We can and probably should.  1643247 has a lot of changes that may deem to be unnecessary and this one line change has the same outcome.  There is no need for the accuracy that the other CL fixes by anticipating the latency time and adding that into the expected wakeup time.
> > 
> > ??? what do you mean by "anticipating" ? there's no anticipating. they are measured so it's always there. What we are trying to handle in this file(power_mgt.c) is about added latency by other interrupts.
> 
> So what I'm saying is 
> what we are handling here is the added latency that is only for the cases like multiple interrupts happens at the same time.

and you are right, there is some alpha that we are adding to the time that we said to wakeup at. The case that we see this is when we do the math from a signed to an unsigned and the "already expired" turns into a very long time before expiring.  If the code changes, latency possibly increased, then "already expired" is still the same meaning, as far as we are concerned with the power level that we should be currently running at.  So we are adding complexity to show we can get closer with the current code, but it does not change anything and the one line of simple adjustment for "already expired" handles the current case and anything that may change in the future.

Patch-set: 2
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>