Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Sorry if this is showing my newness to the EC code... why doesn't the hwtimer present a 64 bit time value and how that is implemented can be contained within hwtimer.c?  If hardware only has 32 bit timers then the extra rollover timer and adjustments when the rollover should have happened but has not yet expressed itself on the RAM portion of the 64 bit work would be contained to the single hardware specific file and getting a timer value would always present a valid 64 bit value.
> 
> This seems cleaner than propagating code around the code base to handle the rollover that may not have adjusted the software portion of the 64 bit value.  The implementation would be abstracted and hidden in a single chip specific file.

Yes, this is the way it should be.

By adding this config option, we can refactor the other chips to be this way chip-by-chip, and eventually remove the config option and the clutter from the assumption that all timers are 32-bit.

Patch-set: 1
Reviewer: Gerrit User 1330845 <1330845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=0
Label: Auto-Submit=0