Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Sorry if this is showing my newness to the EC code... why doesn't the hwtimer present a 64 bit time value and how that is implemented can be contained within hwtimer.c?  If hardware only has 32 bit timers then the extra rollover timer and adjustments when the rollover should have happened but has not yet expressed itself on the RAM portion of the 64 bit work would be contained to the single hardware specific file and getting a timer value would always present a valid 64 bit value.
> > > 
> > > This seems cleaner than propagating code around the code base to handle the rollover that may not have adjusted the software portion of the 64 bit value.  The implementation would be abstracted and hidden in a single chip specific file.
> > 
> > Yes, this is the way it should be.
> > 
> > By adding this config option, we can refactor the other chips to be this way chip-by-chip, and eventually remove the config option and the clutter from the assumption that all timers are 32-bit.
> 
> Honestly, as I mentioned in the but, I also wanted to have 64 bit timer support but it should be done in more clean way.

I agree a more clean manner to handle this so that common code just sees a 64 bit timer, however it is implemented, sounds like a great idea.  I can understand Jack wanting to use a config option so we can refactor one chip at a time and not have an out of control CL and agree this is a fine path with the understanding that at some point the config option would be able to be removed from the general code, if not completely, and all chips should be using the abstracted 64 bit timer.

Patch-set: 1
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>