Update patch set 6
Patch Set 6:

One side-effect of using this CONFIG_* option is a "high-power" charger may appear as a "low power" charger to the user which could be confusing. I suppose if you prevent the AP from booting entirely, then you could prevent this. It might be worth putting this warning in a comment somewhere...

Patch-set: 6
CC: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>