chip/mt_scp/hrtimer: Switch hrtimer to use ULPOSC1/8
To be able to use the SCP in S3, we cannot rely on the 26Mhz clock,
as it's off in S3. Switch to using ULPOSC1 divided by 8 instead.

Also, refactor the code to make it easier for hrtimer to use
ULPOSC1 frequency, and make sure the frequency is a multiple of
8 for the timer to be accurate.

BRANCH=none
BUG=b:134035444
TEST=make BOARD=kukui_scp -j && \
  bash board/kukui_scp/update_scp $IP
  powerd_dbus_suspend

Change-Id: I048431cc062040caea70a5de3709ddd33550439c
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>