driver:lis2mdl: Remove todo
After some digging around I've found that the issue wasn't in the
call to init_mag_cal. Instead it was the fact that we weren't
resetting the chip using the LIS2MDL_FLAG_SW_RESET flag every time
we reset the calibration. Instead of resetting every time, it
actually improves performance to ignore no-op calls here. So the
comment can be resolved.

BUG=b:130417518
TEST=Added reset flag to set ODR command and remove early bail.
Everything works.
BRANCH=None

Change-Id: Ida4afefc4aad0630bb60a4e25371aa050f0886d0
Signed-off-by: Yuval Peress <peress@chromium.org>