ISH: fix UART output error
Currently when there're a lot of prints to UART we don't see all
of the prints. For example,

-------- code --------
for (i = 0, v = 0x000FFFF; i < 1000; i++, v++)
    CPRINTF(">> UART TEST [%d] = 0x%08x\n", i, v);
----------------------

-------- output --------
>> UART TEST [0] = 0x0000ffff
>> UART TEST [1] = 0x00010000
    :
    :
>> UART TEST [56] = 0x00010037
>> UART TEST [57] = 0x00010038
>> UART TEST>>>>>>>>>>>>>>[0.407104 HC 0x400b err 1] <<= print error
------------------------

It only happens with lots of prints and, based on the output, it
looks like we are overwriting data before it's sent out. So This
patch changes the bit used to see if transmitter is ready for new
data from DTRQ to TEMT.

BUG=none
BRANCH=none
TEST=Ran the code above and verified the output on Arcara platform.

Change-Id: I21e9209510cade39f95e161f23ee373007b90e50