Update patch set 3
Patch Set 3: Verified+1

I spent some time testing the original version, it turns out that it takes a long while for the deferred function to get to run, if there are other deferrals waiting.

This was resulting in interruption of the USB stream in case someone runs say 'help' command on the Cr50 console. The deferred function got to run after 81 ms.

The second version adds a check of how much data is already in the buffer when some more UART bytes arrive, and if the buffer is half full or more, the deferred function is cancelled and the interrupt processing proceeds.

With this tweak two full chargen streams from AP and EC keep running without any losses and the rest of tasks still get a fair amount of time to run.

Patch-set: 3
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1