[WIP] I2C_SPI_Slave_Addr in EC code
Patchset1 is the base change 8bit I2C addresses to be 7bit in EC.

Compare that with the current Patchset. This gives an idea where I
would like to go with the I2C/SPI slave addressing.  Let me know
if there are concerns

The reasoning for this has to do with the multiple ways the I2C
and SPI addresses are used and the flag bits being in different
areas.  Doing this will make all of these addresses common

BUG=chromium:971296
BRANCH=none
TEST=make buildall
TEST=verify sensor usage on arcada

Change-Id: I99f58f71ac0c0d85c3475059595c5f9bd28166fe
Signed-off-by: Denis Brockus <dbrockus@chromium.org>