[WIP] I2C_SPI_Slave_Addr in EC code
The reasoning for this has to do with the multiple ways the I2C
and SPI addresses are used and the flag bits being in different
areas.  Doing this will make all of these addresses common

NOTE: Next step...
I added the port and i2c_addr into one structure but the added
code to create this from constants in the many places that it is
created on the fly caused some builds to be too large. So I am
removing port from the structure and trying to reduce the size
so this will no longer be in the way

BUG=chromium:971296
BRANCH=none
TEST=make buildall
TEST=verify sensor usage on arcada

Change-Id: I99f58f71ac0c0d85c3475059595c5f9bd28166fe
Signed-off-by: Denis Brockus <dbrockus@chromium.org>